orientation-event/device-orientation-events-unavailable-on-insecure-origins.html

<!DOCTYPE html>
<title>Device Sensor Events not exposed to insecure origins</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/orientation-event-helpers.js"></script>
<script>

if (window.location.origin != get_host_info().HTTP_ORIGIN) {
  window.location = get_host_info().HTTP_ORIGIN + window.location.pathname;
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
} else {
  test(() => {
    assert_false('DeviceMotionEvent' in window);
    assert_false('DeviceOrientationEvent' in window);
    assert_false('DeviceOrientationAbsoluteEvent' in window);
    assert_false('DeviceMotionEventAcceleration' in window);
    assert_false('DeviceMotionEventRotationRate' in window);
    assert_false('ondevicemotion' in window);
    assert_false('ondeviceorientation' in window);
    assert_false('ondeviceorientationabsolute' in window);
  }, 'Event interfaces and event handlers are not exposed on `window`.');

  promise_test(async (t) => {
    window.ondevicemotion = t.unreached_func("devicemotion event should not be fired.");

    await new Promise(r => t.step_timeout(r, 1000));
  }, 'addEventListener() for `devicemotion` does not crash but the handler never fires.');

  promise_test(async (t) => {
    window.ondeviceorientation = t.unreached_func("deviceorientation event should not be fired.");

    await new Promise(r => t.step_timeout(r, 1000));
  }, 'addEventListener() for `deviceorientation` does not crash but the handler never fires.');

  promise_test(async (t) => {
    window.ondeviceorientationabsolute = t.unreached_func("deviceorientationabsolute event should not be fired.");

    await new Promise(r => t.step_timeout(r, 1000));
  }, 'addEventListener() for `deviceorientationabsolute` does not crash but the handler never fires.');
}
</script>
